Add python logging
It is streaming logs to sys.stdout.

Looks like it works OK with docker logs #CONTAINER, but it will not
generate files. I could not make it work with file and stdout for docker
logs. For now it is based on docker so I will not waste more time on it.

Closes #8